requirements.txt
httpx
psycopg2-binary
APScheduler
fastapi
uvicorn